add plan-map subcomponent
