+++ src/interface/src/app/app.module.ts
@@ -37,6 +37,7 @@ import { SharedModule } from './shared/shared.module';
 import { SignupComponent } from './signup/signup.component';
 import { StringifyMapConfigPipe } from './stringify-map-config.pipe';
 import { TopBarComponent } from './top-bar/top-bar.component';
+import { PlanMapComponent } from './plan/plan-map/plan-map.component';
 
 @NgModule({
   declarations: [
@@ -54,6 +55,7 @@ import { TopBarComponent } from './top-bar/top-bar.component';
     PlanCreateDialogComponent,
     LayerInfoCardComponent,
     PlanComponent,
+    PlanMapComponent,
   ],
   imports: [
     AppRoutingModule,
