+++ src/interface/src/app/plan/plan-map/plan-map.component.html
@@ -1,0 +1,4 @@
+<div class="plan-map-view">
+  <div id="map" class="plan-map"></div>
+  <button mat-raised-button class="expand-map-button" (click)="expandMap()">EXPAND MAP</button>
+</div>
