+++ src/interface/src/app/plan/plan.component.scss
@@ -22,25 +22,6 @@
   height: 100%;
 }
 
-.plan-map-view {
-  background-color: #e9e9e9;
-  flex-grow: 2;
-  position: relative;
-}
-
-.plan-map {
-  height: 100%;
-  width: 100%;
-  z-index: 1;
-}
-
-.expand-map-button {
-  bottom: 12px;
-  position: absolute;
-  right: 12px;
-  z-index: 2;
-}
-
 .plan-scenario-panel {
   background-color: white;
   display: flex;
